v2.14.5

+++ installer/windows/uninstaller/copy_and_run.h
@@ -4,6 +4,6 @@
 
 namespace CopyAndRun
 {
-    int runFirstPhase(const std::wstring &uninstExeFile, const char *lpszCmdParam);
+    unsigned long runFirstPhase(const std::wstring &uninstExeFile, const char *lpszCmdParam);
 };
